Crate trust_dns_server

Trust-DNS is intended to be a fully compliant domain name server and client library.

Goals

Only safe Rust

All errors handled

Simple to manage servers

High level abstraction for clients

Secure dynamic update

New features for securing public information

Re-exports

pub extern crate trust_dns_proto as proto;

Modules

authority
Module for Catalog of Authority zones which are responsible for storing RRSet records.

config
Configuration module for the server binary, named.

error
All defined errors for Trust-DNS

logger
Default logger configuration for the project

server
Server component for hosting a domain name servers operations.

store
All persistent store implementations

Structs

ServerFuture
A Futures based implementation of a DNS server

Functions

version
Returns the current version of Trust-DNS