Module trust_dns_server::authority

Module for Catalog of Authority zones which are responsible for storing RRSet records.

Structs

AnyRecords
An iterator over an ANY query for Records.

BoxedLookupFuture
A boxed lookup future

Catalog
Set of authorities, zones, available to this server.

MessageRequest
A Message which captures the data from an inbound request

MessageResponse
A EncodableMessage with borrowed data for Responses in the Server

MessageResponseBuilder
A builder for MessageResponses

Queries
A set of Queries with the associated serialized data

Enums

AuthLookup
The result of a lookup on an Authority

AuthLookupIter
An iterator over an Authority Lookup

LookupError
A query could not be fullfilled

LookupRecords
The result of a lookup

LookupRecordsIter
Iteratof over lookup records

ZoneType
The type of zone stored in a Catalog

Traits

Authority
Authority implementations can be used with a Catalog

AuthorityObject
An Object safe Authority

LookupObject
An Object Safe Lookup for Authority

UpdateRequest
A type which represents an MessageRequest for dynamic Update.

Type Definitions

LookupResult
Result of a Lookup in the Catalog and Authority

UpdateResult
Result of an Update operation