Enum trust_dns_server::authority::AuthLookup

pub enum AuthLookup {
    Empty,
    Records {
        answers: LookupRecords,
        additionals: Option<LookupRecords>,
    },
    SOA(LookupRecords),
    AXFR {
        start_soa: LookupRecords,
        records: LookupRecords,
        end_soa: LookupRecords,
    },
}

The result of a lookup on an Authority

Lifetimes

'c - the catalogue lifetime

'r - the recordset lifetime, subset of 'c

'q - the queries lifetime

Variants
Empty

No records

Records

Records

Fields of Records

answers: LookupRecords

Authoritative answers

additionals: Option<LookupRecords>

Optional set of LookupRecords

SOA(LookupRecords)

Soa only differs from Records in that the lifetime on the name is from the authority, and not the query

AXFR

An axfr starts with soa, chained to all the records, then another soa...

Fields of AXFR

start_soa: LookupRecords

The first SOA record in an AXFR response

records: LookupRecords

The records to return

end_soa: LookupRecords

The last SOA record of an AXFR (matches the first)

Methods

impl AuthLookup

pub fn answers(
answers: LookupRecords,
additionals: Option<LookupRecords>
) -> Self

Construct and ansert with additional section

pub fn is_empty(&self) -> bool

Returns true if either the associated Records are empty, or this is a NameExists or NxDomain

pub fn was_empty(&self) -> bool

This is an NxDomain or NameExists, and has no associated records

this consumes the iterator, and verifies it is empty

Important traits for AuthLookupIter<'r>
impl<'r> Iterator for AuthLookupIter<'r> type Item = &'r Record;

pub fn iter(&self) -> AuthLookupIter

Conversion to an iterator

pub fn unwrap_records(self) -> LookupRecords

Does not panic, but will return no records if it is not of that type

pub fn take_additionals(&mut self) -> Option<LookupRecords>

Takes the additional records, leaving behind None