Struct trust_dns_server::authority::MessageRequest

pub struct MessageRequest { /* fields omitted */ }

A Message which captures the data from an inbound request

Methods

impl MessageRequest

pub fn id(&self) -> u16

see Header::id()

pub fn message_type(&self) -> MessageType

see Header::message_type()

pub fn op_code(&self) -> OpCode

see Header::op_code()

pub fn authoritative(&self) -> bool

see Header::authoritative()

pub fn truncated(&self) -> bool

see Header::truncated()

pub fn recursion_desired(&self) -> bool

see Header::recursion_desired()

pub fn recursion_available(&self) -> bool

see Header::recursion_available()

pub fn authentic_data(&self) -> bool

see Header::authentic_data()

pub fn checking_disabled(&self) -> bool

see Header::checking_disabled()

pub fn response_code(&self) -> ResponseCode

Return value

The ResponseCode, if this is an EDNS message then this will join the section from the OPT record to create the EDNS ResponseCode

pub fn queries(&self) -> &[LowerQuery]

Question        Carries the query name and other query parameters.

pub fn answers(&self) -> &[Record]

Answer          Carries RRs which directly answer the query.

pub fn name_servers(&self) -> &[Record]

Authority       Carries RRs which describe other authoritative servers.
                May optionally carry the SOA RR for the authoritative
                data in the answer section.

pub fn additionals(&self) -> &[Record]

Additional      Carries RRs which may be helpful in using the RRs in the
                other sections.

pub fn edns(&self) -> Option<&Edns>

RFC 6891, EDNS(0) Extensions, April 2013

6.1.1.  Basic Elements

 An OPT pseudo-RR (sometimes called a meta-RR) MAY be added to the
 additional data section of a request.

 The OPT RR has RR type 41.

 If an OPT record is present in a received request, compliant
 responders MUST include an OPT record in their respective responses.

 An OPT record does not carry any DNS data.  It is used only to
 contain control information pertaining to the question-and-answer
 sequence of a specific transaction.  OPT RRs MUST NOT be cached,
 forwarded, or stored in or loaded from master files.

 The OPT RR MAY be placed anywhere within the additional data section.
 When an OPT RR is included within any DNS message, it MUST be the
 only OPT RR in that message.  If a query message with more than one
 OPT RR is received, a FORMERR (RCODE=1) MUST be returned.  The
 placement flexibility for the OPT RR does not override the need for
 the TSIG or SIG(0) RRs to be the last in the additional section
 whenever they are present.

Return value

Returns the EDNS record if it was found in the additional section.

pub fn sig0(&self) -> &[Record]

Any SIG0 records for signed messages

pub fn max_payload(&self) -> u16

Return value

the max payload value as it's defined in the EDNS section.

pub fn version(&self) -> u8

Return value

the version as defined in the EDNS record

pub fn raw_queries(&self) -> &Queries

Returns the queries passed recieved from the client