Trait trust_dns_server::server::RequestHandler

pub trait RequestHandler: Send + 'static {
    type ResponseFuture: Future<Item = (), Error = ()> + Send + 'static;
    fn handle_request<R: ResponseHandler>(
        &self, 
        request: Request, 
        response_handle: R
    ) -> Self::ResponseFuture;
}

Trait for handling incoming requests, and providing a message response.

Associated Types

type ResponseFuture: Future<Item = (), Error = ()> + Send + 'static

A future for execution of the request

Loading content...
Required methods

fn handle_request<R: ResponseHandler>(
&self,
request: Request,
response_handle: R
) -> Self::ResponseFuture

Determine's what needs to happen given the type of request, i.e. Query or Update.

Arguments

request - the requested action to perform.

response_handle - handle to which a return message should be sent

Loading content...