Struct trust_dns_server::store::forwarder::ForwardAuthority

pub struct ForwardAuthority { /* fields omitted */ }

An authority that will forward resolutions to upstream resolvers.

This uses the trust-dns-resolver for resolving requests.

Methods

impl ForwardAuthority

pub fn try_from_config(
origin: Name,
_zone_type: ZoneType,
config: &ForwardConfig
) -> Result<(Self, impl Future<Item = (), Error = ()>), String>

Read the Authority for the origin from the specified configuration