Struct trust_dns_server::store::in_memory::InMemoryAuthority

pub struct InMemoryAuthority { /* fields omitted */ }

InMemoryAuthority is responsible for storing the resource records for a particular zone.

Authorities default to DNSClass IN. The ZoneType specifies if this should be treated as the start of authority for the zone, is a slave, or a cached zone.

Methods

impl InMemoryAuthority

pub fn new(
origin: Name,
records: BTreeMap<RrKey, RecordSet>,
zone_type: ZoneType,
allow_axfr: bool
) -> Result<Self, String>

Creates a new Authority.

Arguments

origin - The zone Name being created, this should match that of the RecordType::SOA record.

records - The map of the initial set of records in the zone.

zone_type - The type of zone, i.e. is this authoritative?

allow_update - If true, then this zone accepts dynamic updates.

is_dnssec_enabled - If true, then the zone will sign the zone with all registered keys, (see add_zone_signing_key())

Return value

The new Authority.

pub fn empty(origin: Name, zone_type: ZoneType, allow_axfr: bool) -> Self

Creates an empty Authority

Warning

This is an invalid zone, SOA must be added

pub fn clear(&mut self)

Clears all records (including SOA, etc)

pub fn class(&self) -> DNSClass

Get the DNSClass of the zone

pub fn set_allow_axfr(&mut self, allow_axfr: bool)

Enables AXFRs of all the zones records

pub fn secure_keys(&self) -> &[Signer]

Retrieve the Signer, which contains the private keys, for this zone

pub fn records(&self) -> &BTreeMap<RrKey, Arc<RecordSet>>

Get all the records

pub fn records_mut(&mut self) -> &mut BTreeMap<RrKey, Arc<RecordSet>>

Get a mutable reference to the records

pub fn minimum_ttl(&self) -> u32

Returns the minimum ttl (as used in the SOA record)

pub fn serial(&self) -> u32

get the current serial number for the zone.

pub fn upsert(&mut self, record: Record, serial: u32) -> bool

Inserts or updates a Record depending on it's existence in the authority.

Guarantees that SOA, CNAME only has one record, will implicitly update if they already exist.

Arguments

record - The Record to be inserted or updated.

serial - Current serial number to be recorded against updates.

Return value

true if the value was inserted, false otherwise

pub fn secure_zone(&mut self) -> DnsSecResult<()>

(Re)generates the nsec records, increments the serial number nad signs the zone