Struct trust_dns_server::store::sqlite::persistence::Journal

pub struct Journal { /* fields omitted */ }

The Journal is the audit log of all changes to a zone after initial creation.

Methods

impl Journal

pub fn new(conn: Connection) -> PersistenceResult<Journal>

Constructs a new Journal, attaching to the specified Sqlite Connection

pub fn from_file(journal_file: &Path) -> PersistenceResult<Journal>

Constructs a new Journal opening a Sqlite connection to the file at the specified path

pub fn conn(&self) -> MutexGuard<Connection>

Returns a reference to the Sqlite Connection

pub fn schema_version(&self) -> i64

Returns the current schema version of the journal

Important traits for JournalIter<'j>
impl<'j> Iterator for JournalIter<'j> type Item = Record;

pub fn iter(&self) -> JournalIter

this returns an iterator from the beginning of time, to be used to recreate an authority

pub fn insert_record(
&self,
soa_serial: u32,
record: &Record
) -> PersistenceResult<()>

Inserts a record, this is an append only operation.

Records should never be posthumously modified. The message will be serialized into the. the first message serialized to the journal, should be a single AXFR of the entire zone, this will be used as a starting point to reconstruct the zone.

Argument

record - will be serialized into the journal

pub fn insert_records(
&self,
soa_serial: u32,
records: &[Record]
) -> PersistenceResult<()>

Inserts a set of records into the Journal, a convenience method for insert_record

pub fn select_record(
&self,
row_id: i64
) -> PersistenceResult<Option<(i64, Record)>>

Selects a record from the given row_id.

This allows for the entire set of records to be iterated through, by starting at 0, and incrementing each subsequent row.

Arguments

row_id - the row_id can either be exact, or start at 0 to get the earliest row in the list.

pub fn select_schema_version(conn: &Connection) -> PersistenceResult<i64>

selects the current schema version of the journal DB, returns -1 if there is no schema

Arguments

conn - db connection to use

pub fn schema_up(&mut self) -> PersistenceResult<i64>

initilizes the schema for the Journal